+++ js/index.js
@@ -2,17 +2,16 @@ import 'https://cdn.kernvalley.us/js/std-js/deprefixer.js';
 import 'https://cdn.kernvalley.us/js/std-js/shims.js';
 import 'https://cdn.kernvalley.us/js/std-js/theme-cookie.js';
 import 'https://cdn.kernvalley.us/components/share-button.js';
-import 'https://cdn.kernvalley.us/components/share-to-button/share-to-button.js';
 import 'https://cdn.kernvalley.us/components/github/user.js';
 import 'https://cdn.kernvalley.us/components/current-year.js';
 import 'https://cdn.kernvalley.us/components/leaflet/map.js';
 import 'https://cdn.kernvalley.us/components/leaflet/marker.js';
 import 'https://cdn.kernvalley.us/components/pwa/install.js';
 import 'https://cdn.kernvalley.us/components/ad/block.js';
+import 'https://cdn.kernvalley.us/components/weather/current.js';
 import 'https://cdn.kernvalley.us/components/app/list-button.js';
 import { $, ready } from 'https://cdn.kernvalley.us/js/std-js/functions.js';
 import { init } from 'https://cdn.kernvalley.us/js/std-js/data-handlers.js';
-import { HTMLNotificationElement } from 'https://cdn.kernvalley.us/components/notification/html-notification.js';
 import { importGa, externalHandler, telHandler, mailtoHandler } from 'https://cdn.kernvalley.us/js/std-js/google-analytics.js';
 import { registerMapSearch } from './functions.js';
 import { hashChange, stateHandler } from './handlers.js';
@@ -52,73 +51,13 @@ Promise.allSettled([
 		requestIdleCallback(async () => {
 			registerMapSearch();
 			window.addEventListener('popstate', stateHandler);
-
-			if (! await cookieStore.get({ name: 'notified' })) {
-				const notification = new HTMLNotificationElement('Under Construction', {
-					icon: '/img/favicon.svg',
-					body: 'Kern Valley Camping is still under construction',
-					image: 'https://i.imgur.com/6xZrS3mm.jpg',
-					tag: 'construction',
-					dir: 'ltr',
-					lang: 'en',
-					vibrate: 0,
-					requireInteraction: true,
-					data: {
-						share: {
-							title: 'Kern Valley Camping',
-							text: 'Map of camping sites in and around the Kern River Valley',
-							url: location.origin,
-						},
-						home: {
-							url: 'https://kernvalley.us',
-						}
-					},
-					actions: [{
-						title: 'Share',
-						action: 'share',
-						icon: '/img/adwaita-icons/places/folder-publicshare.svg',
-					}, {
-						title: 'Dismiss',
-						action: 'close',
-						icon: '/img/octicons/x.svg',
-					}]
-				});
-
-				notification.addEventListener('notificationclick', ({ action, notification }) => {
-					switch (action) {
-						case 'close':
-							notification.close();
-							break;
-
-						case 'share':
-							if (navigator.canShare({ title: document.title, url: location.href })) {
-								const { title, text, url } = notification.data.share;
-								navigator.share({ title, url, text });
-							}
-							break;
-						case 'home':
-							location.href = notification.data.home.url;
-							break;
-					}
-				});
-
-				notification.addEventListener('close', () => {
-					cookieStore.set({
-						name: 'notified',
-						value: 'yes',
-						path: '/',
-						sameSite: 'strict',
-						secure: true,
-						expires: new Date('2021-01-01'),
-					});
-				});
-			}
 		});
 
 		Promise.all([
 			customElements.whenDefined('leaflet-map'),
 			customElements.whenDefined('leaflet-marker'),
 		]).then(async () => {
+			hashChange();
 			if (history.state === null && location.hash !== '') {
 				if (location.hash.includes(',')) {
 					const [latitude = NaN, longitude = NaN] = location.hash.substr(1).split(',', 2).map(parseFloat);
@@ -164,9 +103,5 @@ Promise.allSettled([
 				uuid: target.id,
 			}, document.title, url.href);
 		}
-	});
-
-	Promise.all(['leaflet-map', 'leaflet-marker'].map(tag => customElements.whenDefined(tag))).then(() => {
-		hashChange();
 	});
 });