+++ sw-config.js
@@ -6,30 +6,28 @@ layout: null
 const config = {
 	version: '{{ site.data.app.version | default: site.version }}',
 	fresh: [
+		'/',
 		'https://apps.kernvalley.us/apps.json',
-		/* Root document */
-		'{{ site.pages | where: "pinned", true | map: "url" | join: "', '" }}',
-		'{{ site.posts | where: "pinned", true | map: "url" | join: "', '" }}',
 	].map(path => new URL(path, location.origin).href),
 	stale: [
 		'/css/index.min.css',
 		'/js/index.min.js',
 		'/img/icons.svg',
+		'https://cdn.kernvalley.us/img/markers.svg',
 		/* Other HTML */
-		'https://cdn.kernvalley.us/components/toast-message.html',
 		'https://cdn.kernvalley.us/components/leaflet/map.html',
-		'https://cdn.kernvalley.us/components/share-to-button/share-to-button.html',
 		'https://cdn.kernvalley.us/components/github/user.html',
+		'https://cdn.kernvalley.us/components/weather/current.html',
+		'https://cdn.kernvalley.us/components/toast-message.html',
 
 		/* JS, `customElements`, etc. */
-		'https://polyfill.io/v3/polyfill.min.js',
 
 		/* CSS */
 		'https://unpkg.com/leaflet@1.7.1/dist/leaflet.css',
 		'https://cdn.kernvalley.us/components/leaflet/map.css',
-		'https://cdn.kernvalley.us/components/toast-message.css',
-		'https://cdn.kernvalley.us/components/share-to-button/share-to-button.css',
 		'https://cdn.kernvalley.us/components/github/user.css',
+		'https://cdn.kernvalley.us/components/weather/current.css',
+		'https://cdn.kernvalley.us/components/toast-message.css',
 
 		/* Images & Icons */
 		'/img/apple-touch-icon.png',
@@ -37,33 +35,19 @@ const config = {
 		'/img/icon-192.png',
 		'/img/icon-32.png',
 		'/img/favicon.svg',
-		'/img/neon.svg',
-		'https://cdn.kernvalley.us/img/adwaita-icons/actions/mail-send.svg',
 		'https://cdn.kernvalley.us/img/adwaita-icons/actions/mark-location.svg',
-		'https://cdn.kernvalley.us/img/octicons/file-media.svg',
 		'https://cdn.kernvalley.us/img/keep-kern-clean.svg',
-
-		/* Social Icons for Web Share API shim */
-		'https://cdn.kernvalley.us/img/octicons/mail.svg',
-		'https://cdn.kernvalley.us/img/logos/facebook.svg',
-		'https://cdn.kernvalley.us/img/logos/twitter.svg',
-		'https://cdn.kernvalley.us/img/logos/google-plus.svg',
-		'https://cdn.kernvalley.us/img/logos/linkedin.svg',
-		'https://cdn.kernvalley.us/img/logos/reddit.svg',
-		'https://cdn.kernvalley.us/img/logos/gmail.svg',
-		'https://cdn.kernvalley.us/img/logos/instagram.svg',
 
 		/* Fonts */
 		'https://cdn.kernvalley.us/fonts/roboto.woff2',
-		'https://cdn.kernvalley.us/fonts/Libertine.woff',
-		'https://cdn.kernvalley.us/fonts/ubuntu.woff2',
 		/* Other */
 	].map(path => new URL(path, location.origin).href),
 	allowed: [
 		/https:\/\/maps\.wikimedia\.org\/osm-intl\/*/,
 		/https:\/\/secure\.gravatar\.com\/avatar\/*/,
 		/https:\/\/i\.imgur\.com\/*/,
-		/https:\/\/avatars2\.githubusercontent\.com\/u\/*/,
+		/https:\/\/*\.githubusercontent\.com\/u\/*/,
 		/https:\/\/api\.github\.com\/users\/*/,
+		/https:\/\/api\.openweathermap\.org\/data\/*/,
 	],
 };